% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{add_suffix}
\alias{add_suffix}
\title{Add suffix}
\usage{
add_suffix(x, y)
}
\arguments{
\item{x}{A character vector.}

\item{y}{A suffix.}
}
\value{
A character vector with upper case version of the input.
}
\description{
Add suffix
}
