% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{FooEnum}
\alias{FooEnum}
\title{A Or B.}
\format{
An object of class \code{FooEnum__bundle} (inherits from \code{savvy_savvyExamples__sealed}) of length 2.
}
\usage{
FooEnum
}
\description{
A Or B.
}
\keyword{datasets}
