% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{or_logical}
\alias{or_logical}
\title{Or operation}
\usage{
or_logical(x, y)
}
\arguments{
\item{x}{A logical vector.}

\item{y}{A logical value.}
}
\value{
A logical vector with filled values (\code{NA} is converted to \code{TRUE}).
}
\description{
Or operation
}
