% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{times_any_int}
\alias{times_any_int}
\title{Multiply Input By Another Input}
\usage{
times_any_int(x, y)
}
\arguments{
\item{x}{An integer vector.}

\item{y}{An integer to multiply.}
}
\value{
An integer vector with values multiplied by \code{y}.
}
\description{
Multiply Input By Another Input
}
