% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{times_any_numeric}
\alias{times_any_numeric}
\title{Multiply Input By Another Input}
\usage{
times_any_numeric(x, y)
}
\arguments{
\item{x}{A real vector.}

\item{y}{A real to multiply.}
}
\value{
A real vector with values multiplied by \code{y}.
}
\description{
Multiply Input By Another Input
}
