% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{times_two_int}
\alias{times_two_int}
\title{Multiply Input By Two}
\usage{
times_two_int(x)
}
\arguments{
\item{x}{An integer vector.}
}
\value{
An integer vector with values multiplied by 2.
}
\description{
Multiply Input By Two
}
