% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{times_two_numeric}
\alias{times_two_numeric}
\title{Multiply Input By Two}
\usage{
times_two_numeric(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with values multiplied by 2.
}
\description{
Multiply Input By Two
}
