% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{to_upper}
\alias{to_upper}
\title{Convert Input To Upper-Case}
\usage{
to_upper(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with upper case version of the input.
}
\description{
Convert Input To Upper-Case
}
