% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\docType{package}
\name{unextendr-package}
\alias{unextendr}
\alias{unextendr-package}
\title{unextendr: Dismantling 'extendr'}
\description{
This is an attempt to explain what extendr does.
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})

}
\keyword{internal}
