% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{to_upper}
\alias{to_upper}
\alias{times_two_int}
\alias{times_two_numeric}
\alias{flip_logical}
\title{Exported Functions}
\usage{
to_upper(x)

times_two_int(x)

times_two_numeric(x)

flip_logical(x)
}
\description{
Exported Functions
}
