% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{flip_logical}
\alias{flip_logical}
\title{Flip Input}
\usage{
flip_logical(x)
}
\arguments{
\item{x}{A logical vector.}
}
\value{
A logical vector with filled values (\code{NA} is converted to \code{TRUE}).
}
\description{
Flip Input
}
