% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\docType{data}
\name{Person}
\alias{Person}
\title{A person with a name}
\format{
An object of class \code{Person__bundle} (inherits from \code{savvy_savvyExamples__sealed}) of length 5.
}
\usage{
Person
}
\description{
A person with a name
}
\keyword{datasets}
