% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{times_any_real}
\alias{times_any_real}
\title{Multiply Input By Another Input}
\usage{
times_any_real(x, y)
}
\arguments{
\item{x}{A real vector.}

\item{y}{A real to multiply.}
}
\value{
A real vector with values multiplied by \code{y}.
}
\description{
Multiply Input By Another Input
}
