% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{to_upper}
\alias{to_upper}
\alias{times_two_int}
\alias{times_two_numeric}
\alias{flip_logical}
\title{Exported Functions}
\usage{
to_upper(x)

times_two_int(x)

times_two_numeric(x)

flip_logical(x)
}
\arguments{
\item{x}{An logical vector.}
}
\value{
A character vector with upper case version of the input.

An integer vector with values multiplied by 2.

A numeric vector with values multiplied by 2.

An logical vector with filled values (\code{NA} is converted to \code{TRUE}).
}
\description{
Exported Functions
}
